# contracts/presence/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/presence/periods": {
      "get": {
        "operationId": "AttendancePeriodController_list",
        "parameters": [
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Periods that have been closed",
        "tags": [
          "Presence — Attendance Periods"
        ]
      }
    },
...
    "/presence/periods/{year}/{month}/closed": {
      "get": {
        "operationId": "AttendancePeriodInternalController_isClosed",
        "parameters": [
          {
            "name": "year",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AttendancePeriodClosedResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid provisioning token"
          }
        },
        "summary": "Whether an attendance month is closed",
        "tags": [
          "Attendance Periods"
        ]
      }
    },
...
    "/presence/scans": {
      "post": {
        "operationId": "ScanController_scan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordScanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Record one scan from a gate device",
        "tags": [
          "Presence — Scans"
        ]
      },
      "get": {
        "operationId": "ScanController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "deviceId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "credentialId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "outcome",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACCEPTED",
                "DUPLICATE",
                "REJECTED_UNKNOWN",
                "REJECTED_REVOKED",
                "REJECTED_INACTIVE",
                "REJECTED_STALE"
              ]
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The scan log, including rejected attempts",
        "tags": [
          "Presence — Scans"
        ]
      }
    },
    "/presence/scans/batch": {
      "post": {
        "operationId": "ScanController_batch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordScanBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Flush scans queued while the device was offline",
        "tags": [
          "Presence — Scans"
        ]
      }
    },
...
  },
  "info": {
    "title": "Presence Service API",
    "description": "Presence Service",
    "version": "0.1.2",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AttendancePeriodClosedDto": {
        "type": "object",
        "properties": {
          "closed": {
            "type": "boolean",
            "description": "Whether the month is closed. Payroll refuses to run against an open month, because a scan recorded afterwards would change a payslip already issued."
          }
        },
        "required": [
          "closed"
        ]
      },
      "AttendancePeriodClosedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AttendancePeriodClosedDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "RecordScanDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "description": "The opaque token read from the card"
          },
          "clientEventId": {
            "type": "string",
            "format": "uuid",
            "description": "Device-generated. Reused on retry, which is what makes a flush idempotent."
          },
          "occurredAt": {
            "type": "string",
            "description": "Server-anchored derivation, sent only for scans taken offline. Absent means \"stamp it now\"."
          },
          "clockAnchorId": {
            "type": "string",
            "format": "uuid",
            "description": "Which anchor produced occurredAt"
          }
        },
        "required": [
          "code",
          "clientEventId"
        ]
      },
      "RecordScanBatchDto": {
        "type": "object",
        "properties": {
          "scans": {
            "description": "An offline flush",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecordScanDto"
            }
          }
        },
        "required": [
          "scans"
        ]
      },
...
    }
  }
}